Exclude other fixups from recent commits
When generating candidates based on most recent commit to each change
file it can easily happen that the last commit is a fixup created
moments ago. This is not very useful as doing a 2nd fixup of the same
original commit will have the same effect and it's more likely to be
shadowing a commit you actually want to fix.

Add a inverted grep for commit messages starting the line with fixup or
squash when calling rev-list