Add --all option
This new option allows to display all the commits in the range
base..HEAD.

This option does not use all the smart closest commit search in the
range. But it allows to use the interactive selection.
It is usefull when the user does a modification unrelated to the target
commit, and knows which commit it is. In that, it may do:
* git commit -m "fixfix"
* git rebase -i base

This new option saves one command, and is shorted to:
* git fixup -A -b base